apps/ml/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-core==2.18.4

numpy==1.26.4
pillow==10.4.0
requests==2.32.3
opencv-python==4.10.0.84
timm==1.0.9
# NOTE: install torch/torchvision separately for Windows (see step 2)
